fixed from PVS-Studio:
V668 There is no sense in testing the 'buffer' pointer against null, as the memory was allocated using the 'new' operator. The exception will be generated in the case of memory allocation error. registry.cpp 68
V668 There is no sense in testing the 'buffer' pointer against null, as the memory was allocated using the 'new' operator. The exception will be generated in the case of memory allocation error. registry.cpp 241
V668 There is no sense in testing the 'm_pszAppName' pointer against null, as the memory was allocated using the 'new' operator. The exception will be generated in the case of memory allocation error. xeventlog.cpp 446